+++ src/Main.java
@@ -5,7 +5,7 @@ import mvp.Presenter;
 import mvp.View;
 import storage.FileStorage;
 import storage.Storage;
-import ui.ConsoleView;
+import ui.ConsoleView;;
 
 public class Main {
     public static void main(String[] args) throws IOException {
@@ -13,6 +13,6 @@ public class Main {
         Model model = new Model(storage);
         View view = new ConsoleView();
         Presenter presenter = new Presenter(view, model);
-        presenter.start();
+        view.start();
     }
 }
