+++ src/mvp/View.java
@@ -1,6 +1,9 @@
 package mvp;
 
 public interface View {
+
+    void start();
+
     void showMainMenu();
 
     void showNoteInputWelcome();
@@ -12,4 +15,6 @@ public interface View {
     String scan();
 
     void print(String outputString);
+
+    void setPresenter(Presenter presenter);
 }
